backend/requirements.txt
# Backend dependencies
flask
flask-cors
flask-caching
python-dotenv
requests
flask-sqlalchemy
flask-bcrypt
pyjwt
mysqlclient
gunicorn
Flask-Migrate
psycopg2-binary

# Machine Learning dependencies
ipykernel
pandas
numpy
matplotlib
seaborn
scikit-learn
jupyter
joblib
openpyxl
imblearn
chardet 
#for running the notebook run this:

# python -m ipykernel install --user --name=myvenv --display-name "Python (myvenv)"
# and select your venv kernal!!

# first lets update pip

# ''' python.exe -m pip install --upgrade pip '''

# # lets install the dependecies

# ''' pip install -r requirements.txt '''
